updated the type of debit and credit columns from unsignedInteger to double

+++ database/migrations/2022_03_29_104430_create_lines_table.php
@@ -16,8 +16,8 @@ return new class () extends Migration {
             $table->id();
             $table->foreignId('account_id')->constrained()->cascadeOnDelete();
             $table->foreignId('transaction_id')->constrained()->cascadeOnDelete();
-            $table->unsignedInteger('debit')->nullable();
-            $table->unsignedInteger('credit')->nullable();
+            $table->double('debit', 12, 2)->nullable();
+            $table->double('credit', 12, 2)->nullable();
             $table->timestamps();
         });
     }
